requirements.txt
torch
torchvision
pandas
numpy
tqdm
pillow
matplotlib
scikit-learn
